pyproject.toml
[project]
name = "shreddit"
version = "0.1.0"
description = "Remove your comment history on Reddit as deleting an account does not do so."
authors = [
    {name = "Hayden", email = "hayden@haydenk.com"},
]
dependencies = [
    "arrow>=1.3.0",
    "praw>=7.7.1",
    "PyYaml>=6.0.1",
    "tornado>=6.4",
    "prometheus-client>=0.20.0",
    "appdirs>=1.4.4",
    "pydantic>=2.6.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "BSD-3-Clause"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
shreddit = {call = "shreddit.app:entry"}
test = "pytest --full-trace --cov-report term-missing --cov=. tests/"
lint = "flake8"
mypy = "mypy --check-untyped-defs ."
black = "black ."

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "flake8>=7.0.0",
    "flake8-bugbear>=24.2.6",
    "black>=24.2.0",
    "bandit>=1.7.7",
]
